chore: revert dev mode, extract recent changes
