--- pages/develop/README.md
@@ -1,9 +1,0 @@
----
-navigation_title: "Develop"
-position: 6
-dev_only: true
----
-
-# Develop
-
-**Development-only tools for maintaining the guide, including merge-based recent changes and a revision overview for every guide page.**
