--- pages/develop/page-revisions/README.md
@@ -1,10 +1,0 @@
----
-navigation_title: "Page revisions"
-position: 2
-dev_only: true
-all_pages_overview: true
----
-
-# Page revisions
-
-**Development overview of all guide pages, sorted by the most recent merge.**
