--- pages/develop/recent-changes/README.md
@@ -1,10 +1,0 @@
----
-navigation_title: "Recent changes"
-position: 1
-dev_only: true
-page_updates_overview: true
----
-
-# Recent changes by merge
-
-**Overview of all merges to main and the guide pages changed in each merge, with pages sorted by number of changed lines.**
