+++ pages/introduction/license/README.md
@@ -1,6 +1,5 @@
 ---
 navigation_title: "License"
-navigation_ignore: true
 position: 4
 ---
