+++ pages/introduction/privacy-policy/README.md
@@ -1,6 +1,5 @@
 ---
 navigation_title: "Privacy Policy"
-navigation_ignore: true
 position: 5
 ---
