--- src/components/content/all-pages/_all-pages-table.scss
@@ -1,77 +1,0 @@
-.all-pages {
-  @include rem(margin-top, $gutter);
-}
-
-.all-pages__intro {
-  margin: 0 0 1em;
-  max-width: 48em;
-}
-
-.all-pages__table-wrap {
-  overflow-x: auto;
-}
-
-.all-pages__table {
-  width: 100%;
-  border-collapse: collapse;
-  font-size: 14px;
-  line-height: 1.5;
-}
-
-.all-pages__table th,
-.all-pages__table td {
-  padding: 0.75em;
-  border-bottom: 1px solid var(--theme-color-medium, $c-gray);
-  text-align: left;
-  vertical-align: top;
-}
-
-.all-pages__table th {
-  font-size: 12px;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-}
-
-.all-pages__date {
-  display: block;
-  margin: 0;
-  font-weight: 600;
-  white-space: nowrap;
-}
-
-.all-pages__section {
-  display: inline-block;
-  margin: 0;
-  padding: 0.1em 0.5em;
-  border-radius: 999px;
-  font-size: 12px;
-  font-weight: 500;
-  line-height: 1.3;
-  letter-spacing: 0.01em;
-  background-color: var(--theme-color-light);
-  color: var(--theme-color-dark);
-  white-space: nowrap;
-}
-
-.all-pages__link {
-  display: inline-block;
-  min-width: 12em;
-}
-
-.all-pages__changes {
-  font-variant-numeric: tabular-nums;
-  white-space: nowrap;
-}
-
-.all-pages__author {
-  display: flex;
-}
-
-.all-pages__empty,
-.all-pages__empty-date,
-.all-pages__empty-changes,
-.all-pages__empty-author {
-  margin: 0;
-  color: var(--theme-color-dark);
-}
