--- src/components/content/page-updates/_page-updates-table.scss
@@ -1,161 +1,0 @@
-.page-updates {
-  @include rem(margin-top, $gutter);
-}
-
-.page-updates__intro {
-  margin: 0 0 1em;
-  max-width: 48em;
-}
-
-.page-updates__list {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  display: grid;
-  gap: 1em;
-}
-
-.page-updates__item {
-  margin: 0;
-}
-
-.page-updates__merge {
-  padding: 1em;
-  border: 1px solid var(--theme-color-medium, $c-gray);
-  border-radius: 10px;
-  background: var(--theme-main-color-inverse, $c-white);
-  box-shadow: 0 4px 14px 0 rgba(0, 0, 0, 0.08);
-}
-
-.page-updates__meta {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 0.5em 0.75em;
-}
-
-.page-updates__date {
-  margin: 0;
-  font-size: 14px;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-}
-
-.page-updates__commit code {
-  font-size: 0.9em;
-}
-
-.page-updates__message {
-  margin: 0.75em 0 0;
-  font-size: 16px;
-  line-height: 1.5;
-}
-
-.page-updates__pages {
-  margin: 0.75em 0 0;
-  padding: 0;
-  list-style: none;
-  border-top: 1px solid var(--theme-color-medium, $c-gray);
-}
-
-.page-updates__page {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: baseline;
-  gap: 0.5em 0.75em;
-  padding: 0.5em 0;
-  border-bottom: 1px solid var(--theme-color-medium, $c-gray);
-
-  font-size: 12px;
-  &:last-child {
-    border-bottom: 0;
-    padding-bottom: 0;
-  }
-}
-
-.page-updates__page-label {
-  display: inline-flex;
-  margin: 0;
-  cursor: pointer;
-}
-
-.page-updates__page-checkbox {
-  margin: 0;
-}
-
-.page-updates__link {
-  flex: 1 1 auto;
-  min-width: 12em;
-}
-
-.page-updates__changes {
-  margin-left: auto;
-  font-size: 0.85em;
-  font-variant-numeric: tabular-nums;
-  white-space: nowrap;
-}
-
-.page-updates__section {
-  display: inline-block;
-  margin: 0;
-  padding: 0.1em 0.5em;
-  border-radius: 999px;
-  font-size: 12px;
-  font-weight: 500;
-  line-height: 1.3;
-  letter-spacing: 0.01em;
-  background-color: var(--theme-color-light);
-  color: var(--theme-color-dark);
-}
-
-.page-updates__author {
-  display: flex;
-}
-
-.page-updates__empty {
-  margin: 0;
-}
-
-.page-updates__output {
-  margin-top: 1em;
-  padding-top: 1em;
-  border-top: 1px solid var(--theme-color-medium, $c-gray);
-}
-
-.page-updates__output-label {
-  display: block;
-  margin-bottom: 0.5em;
-  font-weight: 600;
-}
-
-.page-updates__textarea {
-  width: 100%;
-  min-height: 8em;
-  padding: 0.75em;
-  border: 1px solid var(--theme-color-medium, $c-gray);
-  border-radius: 6px;
-  font-family:
-    ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono',
-    'Courier New', monospace;
-  font-size: 14px;
-  line-height: 1.5;
-  resize: vertical;
-}
-
-.page-updates__actions {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 0.75em;
-  margin-top: 0.75em;
-}
-
-.page-updates__copy {
-  margin: 0;
-}
-
-.page-updates__status {
-  margin: 0;
-  font-size: 14px;
-}
