--- src/components/content/page-updates/page-updates-table.hbs
@@ -1,108 +1,0 @@
-<section
-  class="page-updates{{#if showConfigurator}} page-updates--configurator js-page-updates-configurator{{/if}}"
-  aria-labelledby="page-updates-caption"
->
-  <h2 id="page-updates-caption" class="visuallyhidden">
-    Merges sorted by date with affected pages
-  </h2>
-
-  {{#if showConfigurator}}
-    <p class="page-updates__intro">
-      Select the pages that should appear in the home page recent changes list. Copy the generated configuration below into <code>config/recent-pages.js</code>, then restart <code>npm run dev</code> or run <code>npm run build</code> so the home page picks up the changes. Author, date, and commit details come from the selected merge.
-    </p>
-  {{/if}}
-
-  {{#if merges}}
-    <ol class="page-updates__list">
-      {{#each merges}}
-        <li class="page-updates__item">
-          <article class="page-updates__merge">
-            <div class="page-updates__meta">
-              <time class="page-updates__date" datetime="{{changed}}">
-                {{ formatDateTime changedTimestamp }}
-              </time>
-              {{#if changedBy}}
-                {{> "content/author/author" name=changedBy gravatarUrl=gravatarUrl className="page-updates__author"}}
-              {{else}}
-                <span class="visuallyhidden">Updated by unknown author</span>
-              {{/if}}
-              {{#if commitUrl}}
-                <a
-                  class="page-updates__commit"
-                  href="{{commitUrl}}"
-                  aria-label="View commit {{commitShortId}} in the repository"
-                >
-                  <code>{{commitShortId}}</code>
-                </a>
-              {{else}}
-                <span class="visuallyhidden">Commit unavailable</span>
-              {{/if}}
-            </div>
-            {{#if commitMessage}}
-              <p class="page-updates__message">{{commitMessage}}</p>
-            {{/if}}
-            <ul class="page-updates__pages" aria-label="Affected pages">
-              {{#each pages}}
-                <li class="page-updates__page">
-                  {{#if ../../showConfigurator}}
-                    <label class="page-updates__page-label">
-                      <input
-                        class="page-updates__page-checkbox"
-                        type="checkbox"
-                        name="recent-page"
-                        value="{{url}}"
-                        data-commit="{{../commitId}}"
-                        {{#if isSelected}}checked{{/if}}
-                      >
-                      <span class="visuallyhidden">Show {{title}} on the home page</span>
-                    </label>
-                  {{/if}}
-                  {{#if sectionTitle}}
-                    <span class="page-updates__section theme theme-{{section}}">
-                      <span class="visuallyhidden">Section: </span>{{sectionTitle}}
-                    </span>
-                  {{/if}}
-                  <a class="page-updates__link" href="/{{url}}">{{title}}</a>
-                  <span
-                    class="page-updates__changes"
-                    aria-label="{{linesAdded}} lines added, {{linesDeleted}} lines deleted"
-                  >
-                    +{{linesAdded}} −{{linesDeleted}}
-                  </span>
-                </li>
-              {{/each}}
-            </ul>
-          </article>
-        </li>
-      {{/each}}
-    </ol>
-  {{else}}
-    <p class="page-updates__empty">No pages with merge history were found.</p>
-  {{/if}}
-
-  {{#if showConfigurator}}
-    <div class="page-updates__output">
-      <label class="page-updates__output-label" for="page-updates-configurator-snippet">
-        Configuration
-      </label>
-      <textarea
-        id="page-updates-configurator-snippet"
-        class="page-updates__textarea"
-        rows="10"
-        readonly
-        aria-describedby="page-updates-configurator-status"
-      ></textarea>
-      <div class="page-updates__actions">
-        <button type="button" class="button page-updates__copy">
-          Copy snippet
-        </button>
-        <p
-          id="page-updates-configurator-status"
-          class="page-updates__status"
-          role="status"
-          aria-live="polite"
-        ></p>
-      </div>
-    </div>
-  {{/if}}
-</section>
